+++ docs/vectors.md
@@ -3,4 +3,10 @@ id: vectors
 title: Vectors and Covariates
 ---
 
-TODO
+```scala mdoc:invisible
+import com.stripe.rainier.core._
+import com.stripe.rainier.plot._
+import Jupyter._
+```
+
+Our imaginary chicken farmer from the [previous section](likelihoods.md) has gotten more sophisticated: she's testing three different types of feed to see which one produces more eggs.
