+++ rainier-test/src/test/scala/com/stripe/rainier/optimizer/OptimizerTest.scala
